Add a few opionionated prebuild variants.
Following the pareto principle we can expect 80% of people looking around
for a variant of Hack to just not like the vertical slashed 0 and the
rounded i. Why not have those variants prebuilt?

Obviously it is too wasteful to compute all permutations of these
variants, but that isn't necessary for most.